requirements.txt
torch>=1.12
torchvision
matplotlib
Pillow
numpy
tqdm
wandb
